+++ src/test/java/domain/coordinate/CoordinateTest.java
@@ -1,11 +1,11 @@
-package coordinate;
+package domain.coordinate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.position.Column;
+import domain.position.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import position.Column;
-import position.Row;
 
 class CoordinateTest {
