+++ src/test/java/domain/position/PositionTest.java
@@ -1,4 +1,4 @@
-package position;
+package domain.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
