+++ src/test/java/view/util/PieceTranslatorTest.java
@@ -2,10 +2,10 @@ package view.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.piece.Color;
+import domain.piece.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import piece.Color;
-import piece.Pawn;
 
 class PieceTranslatorTest {
